% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlayout.R
\name{unlayout}
\alias{unlayout}
\title{Make HTML widgets draggable}
\usage{
unlayout(...)
}
\arguments{
\item{...}{HTML widgets.}
}
\description{
Make HTML widgets draggable
}
\examples{
library(plotly)
p1 <- plot_ly(economics, x = ~ pop, width = 350)
p2 <- plot_ly(economics, x = ~ date, y = ~ pop, height = 240, width = 600)
p3 <- plot_ly(z = ~ volcano)
unlayout(p1, p2, p3)
}
